Simplify qtee_supplicant listener init and deinit framework
Simplified qtee_supplicant implementation by introducing a single callback
pair init() and deinit() for listener initialization and de-initialization.
This re-design ensures that qtee_supplicant should not perform any 'default'
initialization for any listener, such as for time or fs listener. All init
and deinit work is offloaded to the listener implementation.

The CListenerCBO callback object implementation is shared by the time, fs
and gpfs listeners. This CBO is planned to be deprecated eventually as
these listeners move to a client-context from a daemon-context. This change
primarily prepares the listeners for this future transition.

Signed-off-by: Harshal Dev <hdev@qti.qualcomm.com>